requirements.txt
Flask==3.0.3
Jinja2==3.1.4
PyYAML==6.0.2
Werkzeug==3.0.3